refactor: move isCfiInRange function to utils and clean up imports
