+++ apps/web/src/routes/kiwi/-header/bookmark-button.tsx
@@ -1,27 +1,15 @@
 import { Bookmark } from "lucide-react";
 
-import { EpubCFI } from "@bookiwi/epubjs";
 import { useAtomValue, useSetAtom } from "@bookiwi/jotai";
 
 import {
   bookmarksAtom,
-  Cfi,
   currentCfiAtom,
   currentLocationAtom,
   removeBookmarkAtom,
   addBookmarkAtom,
 } from "../-reader/atoms";
-
-const isCfiInRange = (cfi: Cfi, currentCfi: Cfi): boolean => {
-  const cfiInstance = new EpubCFI();
-  const isStartInRange =
-    cfiInstance.compare(cfi.start, currentCfi.start) >= 0 &&
-    cfiInstance.compare(cfi.start, currentCfi.end) <= 0;
-  const isEndInRange =
-    cfiInstance.compare(cfi.end, currentCfi.start) >= 0 &&
-    cfiInstance.compare(cfi.end, currentCfi.end) <= 0;
-  return isStartInRange || isEndInRange;
-};
+import { isCfiInRange } from "../-reader/utils";
 
 function BookmarkButton() {
   const currentCfi = useAtomValue(currentCfiAtom);
