+++ apps/web/src/routes/kiwi/-reader/index.ts
@@ -1,2 +1,1 @@
 export * from "./components";
-export * from "./utils";
