+++ apps/web/src/routes/kiwi/-reader/types/index.ts
@@ -1,0 +1,4 @@
+export interface Cfi {
+  start: string;
+  end: string;
+}
