+++ apps/web/src/routes/kiwi/-reader/utils/cfi.ts
@@ -1,0 +1,14 @@
+import { EpubCFI } from "@bookiwi/epubjs";
+
+import { Cfi } from "../types";
+
+export const isCfiInRange = (cfi: Cfi, currentCfi: Cfi): boolean => {
+  const cfiInstance = new EpubCFI();
+  const isStartInRange =
+    cfiInstance.compare(cfi.start, currentCfi.start) >= 0 &&
+    cfiInstance.compare(cfi.start, currentCfi.end) <= 0;
+  const isEndInRange =
+    cfiInstance.compare(cfi.end, currentCfi.start) >= 0 &&
+    cfiInstance.compare(cfi.end, currentCfi.end) <= 0;
+  return isStartInRange || isEndInRange;
+};
