+++ apps/web/src/routes/kiwi/-reader/utils/index.ts
@@ -2,3 +2,4 @@ export * from "./nav";
 export * from "./styles";
 export * from "./selection";
 export * from "./anchor";
+export * from "./cfi";
